requirements.txt
pyserial
nrfutil